requirements.txt
python-binance>=1.0.16
python-dotenv>=1.0.0
